\name{spssstatistics-package}
\alias{spssstatistics-package}
\alias{spssstatistics}
\docType{package}
\title{
Top level package for driving IBM SPSS Statistics from R
}
\description{
This package loads the latest version of the spss package (installed for the current R version) for use in running IBM® SPSS® Statistics from R. An overview of the features in the spss package is described in the help for \code{\link{spsspkg}}.

Each version of IBM SPSS Statistics has an associated version of the spss package (e.g. spss230 for IBM SPSS Statistics 23). A different version of the spss package, for the current version of R, can be loaded by setting the value of statistics_version in spssstatistics.ini, located in the spssstatistics folder. 
}
\details{
\tabular{ll}{
Package: \tab spssstatistics\cr
Type: \tab Package\cr
Date: \tab 2015-02-10\cr
}
© Copyright IBM Corp. 1989, 2015}

\references{
}
\keyword{ package }
\seealso{\code{\link{spsspkg}}
}
